Added the FadeCanvas prefab to all 3 greybox scenes
Made a change in SceneChangeTrigger so that the canvas will fade in.

Had an small issue with the fading out, throught it would be fixed with the loading code that I added. Didn't fix that, but might be handy for later when the scenes get bigger.

Eventually fixed the issue by adding a small delay to the fading out method/function in the FadeCanvas script