Recycling Across Scenes
Changes in all the scenes, and recycling + scenechangetrigger scripts.

In the scenes, added the recycling prefabs. For the scenes that don't have a recycling bin, I disabled the ability to recycle there. But the script needs to be there.
On every player I added a gameobject where the recycled objects will be instantiated on.
In the sceneChangeTrigger script, i made changes so that it'll move the GameObject with all the recycled onbjects on the players head to the next scene. In that scene, it'll put all the objects that were on the players head on this players head.

This did cause some weird lighting results, but they are not there in a build.