Added prefabs and script
Added a script for the Vacuum cleaner to interact with it: InteractVacuumCleaner.
Added some folders in the Assets. 1 in Interactibles: VacuumCleaner. And 1 in the Assets itself: Vacuum Cleaner. I added a prefab for the interactible vacuum cleaner, THE vacuum cleaner that'll follow you, and 1 that contains the greybox model of the vacuum cleaner