Changes in all the scenes - replaced Player by prefab Player
In all the other scenes, except SampleScene, the player was not the prefab Player. I replaced it and put all the references it needed.
(It can be that I missed some and an error will occur. The error should itself be clear enough on what reference is missing. In samplescene you can also look on what reference is missing since everything (except the progress bar) is there